Trying to get back to baseline
Gitlab-instance dockerfile relies on an image that included the VOLUME directive which messes up the gitlab directories.  Now refers to a different base

Updates the base image definition to refer to the proper sysctl.rb (should be gitlab_sysctl.rb) to keep it from erroring out b/c it can't run in privileged mode.